Completed Order Module

+++ StoreMaster/src/main/java/com/akshay/StoreMaster/dto/CartItemDTO.java
@@ -3,15 +3,17 @@ package com.akshay.StoreMaster.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class CartItemDTO {
     private Long productId;
     private String productName;
     private int quantity;
-    private int price;
+    private BigDecimal price;
 
-    public CartItemDTO(Long productId, String productName, int quantity, int price) {
+    public CartItemDTO(Long productId, String productName, int quantity, BigDecimal price) {
         this.productId = productId;
         this.productName = productName;
         this.quantity = quantity;
